packages/cloud-api/pyproject.toml
[project]
name = "ad-detection-cloud"
version = "0.1.0"
description = "Cloud API for remote fleet management"
authors = [
    {name = "Live Ad Detection Team", email = "dev@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pydantic[email]>=2.4.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.0",  # PostgreSQL driver (sync)
    "asyncpg>=0.29.0",  # PostgreSQL driver (async)
    "pyjwt>=2.8.0",  # JWT tokens
    "bcrypt>=4.1.0",  # Password hashing
    "python-multipart>=0.0.6",  # File uploads
    "httpx>=0.25.0",  # HTTP client
    "redis>=5.0.0",  # Caching and pub/sub
    "celery>=5.3.0",  # Background tasks
    "prometheus-client>=0.18.0",  # Metrics
    "ad-detection-common>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "aiosqlite>=0.19.0",  # Async SQLite for testing
    "black>=23.9.0",
    "ruff>=0.0.292",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
